Add clang-tidy checking to build
make_rules/arm_gcc_rules.mak:
* Separate ARM_FLAGS from CXX_FLAGS, C_FLAGS, AS_FLAGS, LD_FLAGS
  in gcc compiler tool chain.
* Add clang tidy command to compiler command.

make_rules/checker_rules.mak:
  Add clang-tidy definitions and rules.
  Output errors and warning to _checks/<source_file_name>.<ext>.check
To invoke checking: make checks

Use angle-brackets for Nordic SDK header files,
  including softdevice headers.
  Makefile: Use Nordic system files
  Change gcc include directive from -I to -isystem

project_assert.h:
  Use stdlib assert() when not compiling for ARM.
    This eliminates a raft of tidy conditional complaints.
  Rename unit_tests/itest_assert.cc -> assert_stubs.cc
    This file is still compiled in, yet no longer used.
    It hurts nothing, so keep it in case it is useful later.

+++ nordic/peripherals/usart.h
@@ -67,7 +67,7 @@ enum usart_event_type
 struct usart_event_t
 {
     enum usart_event_type   type;
-    uint32_t                value;
+    size_t                  value;
 };
 
 /**
